Changed Eval behaviour and added loadd function
Eval will now run until it cannot discover any executable maps, recursively.
This might slow down execution a bit but it shouldn't matter. This removes
the need to repeat Eval a few times and not being sure if a config
was fully executed.

loadd is now implemented as a function. It will load all files from a directory
specified via the key dir: and with suffix in the key suffix:
The returned data is a executable map with the merge function, which
with the new Eval semantics will automatically expand to the correct
result.

This should enable config writers to easily split configs into independent
files. Be aware that all files are still merged into one keymap so the
fiels must specify the full path into their target if you don't want
values on top level.