Type cast simplification and fix potential bug in unmarshal
This commit should fix a bug in the unmarshal with atomic types and using the UnmarshalSECL interface.

Additionally the =>type function gains the alias `=>T` and now may be called via `(=>type <type> <value>)`
in addition to `(=>type type:<type> <value>)`.
Note that the new syntax is position sensitive, so `(=>type <value> <type>)` does not work.